LayoutTests/media/modern-media-controls/localized-strings/replaced-string-expected.txt

Testing that we can load a replaced string using UIString.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS UIString('Test replaced string') is "Value 15 should be printed here"
PASS successfullyParsed is true

TEST COMPLETE
